[FIX] Add fallthrough comments to satisfy GCC-7.
With GCC 7.3.0 and -Wextra (implying -Wimplicit-fallthroug) we need to
have either an attribute or a comment suggesting that the fallthrough is
intentional.

The attribute is not portable to GCC 6, so it would require an
additional verification of the compiler version, or a non-standard
macro.

The comments, as they were before this commit, should also work, but
they are ineffective due to #endif clause.

This was reported as a bug in GCC
(https://gcc.gnu.org/bugzilla/show_bug.cgi?id=77817), but apparently
it's to complicated to fix. It definitely should be fixed, but to keep
the code simple we'll add the comment after #endif.

Obviously this is dangerous, as it can hide the missing `break` when the
preprocessor condition is not met, but it's probably safer than adding a
not-easily-maintainable macro.

I leave the original comments, so that they suggest that the new one
applies to the #if block.