[#30829] tcg: Update `TCG_COND_*` values and `tcg_gen_movcond_i32`
Parts of this commit (appart from adding `TCG_COND_NEVER` and
`TCG_COND_ALWAYS`) are designed to prepare tlib for future support for
`TCG_TARGET_HAS_movcond_i32`.

Adapted from `include/tcg/tcg.h` (MIT) from
https://github.com/thesofproject/qemu/tree/sof-v5.2.0

An ifdef was added to support native `op_movcond_i32`. It's not yet
supported in tcg-target, but we might want to add it in future.

The `cond_name` array in `tcg.c` was removed as not used.