[#38755] xtensa: Fix PCs
For a proper cooperation with our tlib API, `pc_next` should be `pc`
instead of `npc`. `npc` is an arch-specific field (used in RISC-V and
SPARC) which shouldn't really be in the common `DisasContextBase`.

The `dc->base.pc = dc->base.pc_next` line is removed because now both
these fields are the same.