[DEV] ricv: sparc: Move `npc` from `DisasContextBase` to `DisasContext`
The `npc` is only used internally in RISC-V and SPARC `translate.c` so
it was moved to the arch-specific `DisasContext`.

Having it in the arch-independent `DisasContextBase` made an impression
it's an alternative for QEMU's `pc_next`. It isn't so let's avoid
further mistakes like the one just fixed for Xtensa.