[#34327] host-utils.h: Fix builtin checks
The `__has_builtin` macro was previously problematic in `clrsb32`
because it's been added with GCCv10. The builtins we use were actually
available since GCCv4.7. Using `#if 1` was a WIP change made to quickly
make Xtensa adjustments buildable and it slipped through the MR.

The `clrsb32` function had to be moved after `clz32` after fixing ifdef.
It's called if `__builtin_clrsb` isn't available.

Defining the `__has_builtin` macro for GCC as 1 ensures the builtins
will still be used with older GCCs too.