[#41589] bit_helper.h: Implement ror, rol and ctpop functions
ROR/ROL stands for Rotate Bits Right/Left. The ROR operation move
the 32-bit/64-bit value to the right, which causes some bits to be lost.
The lost bits are then "glued" back on the left side of the registers.
The ROL operation is implemented analogously.

The CTPOP operation returns the number of bits set in the given integer type.

Signed-off-by: Robert Winkler <rwinkler@antmicro.com>