[#41589] arm64: Provide correct CPU state during AArch32 translation
The `cpu_get_tb_cpu_state()` and `cpu_pc_from_tb()` functions are used
to obtain the CPU state during translation handling. The first function
is used when looking for/creating a translation block.
The second is to restore the PC when the EXIT_TB_FORCE flag is encountered.

The function passed incorrect information about the Program Counter (PC)
for the AArch32 state. Besides that, it missed information about
Thumb execution and conditional execution bits.

Signed-off-by: Robert Winkler <rwinkler@antmicro.com>