[#43980] arm64: Add interrupts handling for AArch32
The main skeleton for the function is taken from the
arch/arm/helper.c:do_interrupt() function. Then the function was extended
for handling debug exceptions similarly to AArch64, also an initial
setup for handling Hypervisor mode interrupts was prepared.
Nevertheless, the handler still lacks proper support for Hypervisor traps and
special cases related to Virtualization.

The exception handling is implemented according to the
Arm Architecture Reference Manual for A-profile architecture:
* the IL field is described in G8.2.73 (HSR description)
* return from exception address is set according to G1.12.3
* PSTATE(CPSR) values are set according to G1.12.5
* MOE bits of DBGDSCR are set according to G8.3.14

Co-authored-by: Krzysztof Sychla <ksychla@antmicro.com>
Signed-off-by: Robert Winkler <rwinkler@antmicro.com>