[DEV] arm64: Remove `has_elX` fields
The fields were supposed to be used to restore the initial CPU EL2/3 features after reset
but it makes more sense to keep the EL2/3 features set with `tlib_set_available_els`.

Also, the constraint that CPUs without support for the given EL can't have it enabled
with `tlib_set_available_els` was removed.